move to ReducerBuilder
